+++ resources/js/bootstrap/stores/system.js
@@ -82,6 +82,7 @@ export const useSystemStore = defineStore('system', {
         file: {},
         geometryTypes: [],
         datatypeData: {},
+        accessPoints: {},
     }),
     getters: {
         translateConcept: state => url => {
@@ -117,7 +118,7 @@ export const useSystemStore = defineStore('system', {
     },
     actions: {
         getConceptById(id) {
-            return this.concepts[id];  
+            return this.concepts[id];
         },
         setAppState(state) {
             this.appInitialized = state;
@@ -156,6 +157,7 @@ export const useSystemStore = defineStore('system', {
             this.colorSets = preData.colorSets;
             this.hasAnalysis = preData.analysis;
             this.datatypeData = preData.datatype_data;
+            this.accessPoints = preData.accesspoints;
             entityStore.initializeEntityTypes(preData.entityTypes);
             userStore.setPreferences(preData.preferences);
             // locale.value = this.getPreference('prefs.gui-language');